+++ Note ton STA/src/com/supinfo/notetonsta/DAO/JPA/JPASpeakerDAO.java
@@ -60,7 +60,7 @@ EntityManagerFactory emf;
 	}
 
 	@Override
-	public Speaker getSpeakerById(int id) {
+	public Speaker getSpeakerById(Long id) {
 		EntityManager em = emf.createEntityManager();
 		
 		try {
