+++ Note ton STA/src/com/supinfo/notetonsta/DAO/SpeakerDAO.java
@@ -6,6 +6,6 @@ public interface SpeakerDAO {
 	public void addSpeaker(Speaker speaker);
 	public void updateSpeaker(Speaker speaker);
 	public void deleteSpeaker(Speaker speaker);
-	public Speaker getSpeakerById(int id);
+	public Speaker getSpeakerById(Long id);
 	public boolean authenticateSpeaker(String username,String password);
 }
